style: Fix compiler warnings and apply formatting
- Fix type conversion warnings in multiple game and platform files
- Apply clang-format to affected files
- Update macOS Info.plist version metadata
- Remove debug output from test-ios-headless.sh

Files changed: actions.cc, automap.cc, bmpdlog.cc, combatai.cc, config.cc,
reaction.cc, audio.cc, audiof.cc, assoc.cc, Info.plist, test-ios-headless.sh

+++ src/game/bmpdlog.cc
@@ -1085,7 +1085,7 @@ int save_file_dialog(char* title, char** fileList, char* dest, int fileListLengt
     char fileNameCopy[32];
     strncpy(fileNameCopy, dest, 32);
 
-    int fileNameCopyLength = strlen(fileNameCopy);
+    int fileNameCopyLength = static_cast<int>(strlen(fileNameCopy));
     fileNameCopy[fileNameCopyLength + 1] = '\0';
     fileNameCopy[fileNameCopyLength] = ' ';
 
@@ -1176,7 +1176,7 @@ int save_file_dialog(char* title, char** fileList, char* dest, int fileListLengt
                     }
 
                     fileNameCopy[index] = '\0';
-                    fileNameCopyLength = strlen(fileNameCopy);
+                    fileNameCopyLength = static_cast<int>(strlen(fileNameCopy));
                     fileNameCopy[fileNameCopyLength] = ' ';
                     fileNameCopy[fileNameCopyLength + 1] = '\0';
